Move hbs to its own subfolder
This way alternative template engines can be accommodated in the future.
